Modifications as LFA suggested
