+++ Tests/test_main.py
@@ -1,15 +1,18 @@
+'''Red-Flags Tests file'''
 import unittest
-from main import app, save_redflag, all_redflags
-from main import spec_redflag, del_redflag, edit_redflag
-from models import Incident
+from Routes.main import app, save_redflag, all_redflags
+from Routes.main import spec_redflag, del_redflag, edit_redflag
+from Models.models import Incident
 import json
 
 
 class TestRedflag(unittest.TestCase):
+    '''Class to test Red-Flag'''
     def setUp(self):
         self.test_client = app.test_client()
 
     def test_all_redflag(self):
+        '''Get all Red-Flags test'''
         redflag = {
             "Images": "Images",
             "Videos": "Videos",
@@ -21,7 +24,7 @@ class TestRedflag(unittest.TestCase):
             "status": "draft",
             "type": "RedFlag"
         }
-        self.test_client.post('api/v1/red-flags', json = redflag)
+        self.test_client.post('api/v1/red-flags', json=redflag)
 
         response = self.test_client.get('api/v1/red-flags')
         data = json.loads(response.data)
@@ -29,50 +32,52 @@ class TestRedflag(unittest.TestCase):
         self.assertEqual(len(data['data']), 1)
 
     def test_save_redflag(self):
+        '''Post a Red-Flag Test'''
         redflag1 = {
-        "Images": "Images",
-        "Videos": "Videos",
-        "comment": "Lorem Ipsum is simply dummy text of the century.",
-        "createdBy": "Patrick Kikomeko",
-        "createdOn": "21/12/18",
-        "id": 0,
-        "location": "Kampala",
-        "status": "draft",
-        "type": "RedFlag"
+            "Images": "Images",
+            "Videos": "Videos",
+            "comment": "Lorem Ipsum is simply dummy text of the century.",
+            "createdBy": "Patrick Kikomeko",
+            "createdOn": "21/12/18",
+            "id": 0,
+            "location": "Kampala",
+            "status": "draft",
+            "type": "RedFlag"
         }
         response = self.test_client.post(
             'api/v1/red-flags',
-            json = redflag1)
+            json=redflag1)
         self.assertEqual(response.status_code, 201)
 
     def test_missing_save_redflag(self):
-        #Missing 2 parameters; images and videos
+        '''Post a Red-Flag Test while missing 2 parameters; images and videos'''
         redflag2 = {
-        "comment": "Lorem Ipsum is simply dummy text of the century.",
-        "createdBy": "Patrick Kikomeko",
-        "createdOn": "21/12/18",
-        "id": 0,
-        "location": "Kampala",
-        "status": "draft",
-        "type": "RedFlag"
+            "comment": "Lorem Ipsum is simply dummy text of the century.",
+            "createdBy": "Patrick Kikomeko",
+            "createdOn": "21/12/18",
+            "id": 0,
+            "location": "Kampala",
+            "status": "draft",
+            "type": "RedFlag"
         }
         reponse = self.test_client.post(
             'api/v1/red-flags',
-            json = redflag2
+            json=redflag2
         )
         self.assertEqual(reponse.status_code, 400)
 
 
     def test_none_save_redflag(self):
-        #Empty Input
+        '''Post a Red-Flag test with empty input'''
         redflag3 = {}
         reponse = self.test_client.post(
             'api/v1/red-flags',
-            json = redflag3
+            json=redflag3
         )
         self.assertEqual(reponse.status_code, 400)
 
     def test_spec_redflag(self):
+        '''Get a specific Red-Flag Test'''
         redflag = {
             "Images": "Images",
             "Videos": "Videos",
@@ -84,7 +89,7 @@ class TestRedflag(unittest.TestCase):
             "status": "draft",
             "type": "RedFlag"
         }
-        self.test_client.post('api/v1/red-flags', json = redflag)
+        self.test_client.post('api/v1/red-flags', json=redflag)
         redflag = {
             "Images": "Images",
             "Videos": "Videos",
@@ -96,14 +101,14 @@ class TestRedflag(unittest.TestCase):
             "status": "draft",
             "type": "RedFlag"
         }
-        self.test_client.post('api/v1/red-flags', json = redflag)
+        self.test_client.post('api/v1/red-flags', json=redflag)
         response = self.test_client.get('api/v1/red-flags/0')
         data = json.loads(response.data)
-        print (data)
         self.assertEqual(response.status_code, 200)
 
 
     def test_del_redflag(self):
+        '''Delete a Red-Flag Test'''
         redflag = {
             "Images": "Images",
             "Videos": "Videos",
@@ -115,13 +120,13 @@ class TestRedflag(unittest.TestCase):
             "status": "draft",
             "type": "RedFlag"
         }
-        self.test_client.post('api/v1/red-flags', json = redflag)
+        self.test_client.post('api/v1/red-flags', json=redflag)
         response = self.test_client.delete("/api/v1/red-flags/0")
         data = json.loads(response.data)
-        print(data)
         self.assertEqual(response.status_code, 200)
 
     def test_edit_redflag(self):
+        '''Edit a Red-Flag Test'''
         redflag = {
             "Images": "Images",
             "Videos": "Videos",
@@ -133,7 +138,7 @@ class TestRedflag(unittest.TestCase):
             "status": "draft",
             "type": "RedFlag"
         }
-        self.test_client.post('api/v1/red-flags', json = redflag)
+        self.test_client.post('api/v1/red-flags', json=redflag)
         redflag1 = {
             "Images": "Images1",
             "Videos": "Videos1",
@@ -145,9 +150,6 @@ class TestRedflag(unittest.TestCase):
             "status": "draft",
             "type": "RedFlag"
         }
-        response = self.test_client.patch("/api/v1/red-flags/0", json = redflag1)
+        response = self.test_client.patch("/api/v1/red-flags/0", json=redflag1)
         data = json.loads(response.data)
-        print(data)
         self.assertEqual(response.status_code, 201)
-
-
