+++ run.py
@@ -1,0 +1,4 @@
+from Routes.main import app
+
+if __name__=='__main__':
+    app.run(host="0.0.0.0", port="5000", debug=True)
